+++ src/components/value-list/value-list.tsx
@@ -30,11 +30,12 @@ import {
   removeItem,
   selectSiblings,
   setFocus,
-  setUpItems
+  setUpItems,
+  moveItemIndex
 } from "../pick-list/shared-list-logic";
 import List from "../pick-list/shared-list-render";
-import { getRoundRobinIndex } from "../../utils/array";
 import { createObserver } from "../../utils/observers";
+import { InteractiveComponent, updateHostInteraction } from "../../utils/interactive";
 
 /**
  * @slot - A slot for adding `calcite-value-list-item` elements. Items are displayed as a vertical list.
@@ -47,7 +48,8 @@ import { createObserver } from "../../utils/observers";
 })
 export class ValueList<
   ItemElement extends HTMLCalciteValueListItemElement = HTMLCalciteValueListItemElement
-> {
+> implements InteractiveComponent
+{
   // --------------------------------------------------------------------------
   //
   //  Properties
@@ -136,6 +138,10 @@ export class ValueList<
 
   componentDidLoad(): void {
     this.setUpDragAndDrop();
+  }
+
+  componentDidRender(): void {
+    updateHostInteraction(this);
   }
 
   disconnectedCallback(): void {
@@ -272,9 +278,7 @@ export class ValueList<
     event.preventDefault();
 
     const { el } = this;
-    const moveOffset = event.key === "ArrowDown" ? 1 : -1;
-    const currentIndex = items.indexOf(item);
-    const nextIndex = getRoundRobinIndex(currentIndex + moveOffset, items.length);
+    const nextIndex = moveItemIndex(this, item, event.key === "ArrowUp" ? "up" : "down");
 
     if (nextIndex === items.length - 1) {
       el.appendChild(item);